backend/requirements.txt
Flask
Flask-Cors
python-dotenv
python-dateutil>=2.9.0.post0
supabase
